Tests for summary rendering now invoke the CukedoctorScenarioRenderer rather than CukedoctorConverterImpl. CukedoctorConverter.renderSummary is now deprecated.
